Fixed ICE-6122 - failure to remove unknown files

+++ cs/demo/IceGrid/simple/expect.py
@@ -44,6 +44,7 @@ if Util.isMono():
 
 simple.run('client.exe', desc[0:len(desc)-4])
 
-os.remove("tmp_application.xml")
-os.remove("tmp_application_with_template.xml")
-os.remove("tmp_application_with_replication.xml")
+if Util.isMono():
+    os.remove("tmp_application.xml")
+    os.remove("tmp_application_with_template.xml")
+    os.remove("tmp_application_with_replication.xml")
